api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.120.0",
    "fastapi-clerk-auth>=0.0.7",
    "openai>=2.6.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.38.0",
]
